chore: :building_construction: Relocate Badge component
