+++ libs/dana/src/components/data-display/index.ts
@@ -1,4 +1,3 @@
 export * from './card';
-export * from './badge';
 export * from './table';
 export * from './divider';
